requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
alembic==1.13.1
anyio==4.2.0
asyncpg==0.29.0
black==23.12.1
greenlet==3.0.3
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
ffmpeg-python==0.2.0